Add autotag
This commit adds the autotag script and uses it in the deployment
steps to automate setting and moving release candidate, release,
and prod tags.  The actual steps that perform these tag
manipulations are stubbed out right now, as they are slated to be
done using API calls to a sidecar container that has not yet been
created and will only be deployable from a T-Mobile internal
runner.