Minor cosmetic cleanups
This commit introduces some minor cosmetic cleanups for the .gitlab-ci.yml
file.  First, a variable is introduced referencing the autotag image to
use for the deployment jobs.  Second, an outdated comment was removed.
Third, the docker commands in the package stage have been updated to use
gitlab variables for the registry.